pyproject.toml
[project]
name = "learn-langchain"
version = "0.1.0"
description = "Learning of poetry"
authors = [
    {name = "puggo",email = "puggoo1145@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10, <4.0"

dependencies = [
    "openai (>=1.61.0,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain (>=0.3.17,<0.4.0)",
    "colorama (>=0.4.6,<0.5.0)"
]
[tool.poetry.group.dev.dependencies]
mypy = "^1.14.1"

[tool.poetry]
package-mode = false

[tool.poetry.scripts]
typecheck = "mypy src/app.py"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
